+++ src/components/HOCs/WithReviewTasks/WithReviewTasks.js
@@ -3,8 +3,10 @@ import { connect } from 'react-redux'
 import _omit from 'lodash/omit'
 import _get from 'lodash/get'
 import _debounce from 'lodash/debounce'
-import { fetchReviewNeededTasks } from '../../../services/Task/TaskReview/TaskReviewNeeded'
-import { fetchReviewedTasks } from '../../../services/Task/TaskReview/TaskReviewed'
+import { fetchReviewNeededTasks }
+       from '../../../services/Task/TaskReview/TaskReviewNeeded'
+import { fetchReviewedTasks }
+       from '../../../services/Task/TaskReview/TaskReviewed'
 
 const DEFAULT_PAGE_SIZE = 5
 
@@ -73,13 +75,13 @@ const mapStateToProps = state => ({
 
 const mapDispatchToProps = (dispatch, ownProps) => ({
   updateReviewNeededTasks: (searchCriteria={}, pageSize=DEFAULT_PAGE_SIZE) => {
-    dispatch(fetchReviewNeededTasks(searchCriteria, pageSize))
+    return dispatch(fetchReviewNeededTasks(searchCriteria, pageSize))
   },
   updateReviewedTasks: (searchCriteria={}, pageSize=DEFAULT_PAGE_SIZE) => {
-    dispatch(fetchReviewedTasks(searchCriteria, false, pageSize))
+    return dispatch(fetchReviewedTasks(searchCriteria, false, pageSize))
   },
   updateUserReviewedTasks: (searchCriteria={}, pageSize=DEFAULT_PAGE_SIZE) => {
-    dispatch(fetchReviewedTasks(searchCriteria, true, pageSize))
+    return dispatch(fetchReviewedTasks(searchCriteria, true, pageSize))
   },
 })
