--- src/components/ReviewTasksDashboard/ReviewTasksDashboard.scss
@@ -1,202 +1,0 @@
-@import 'variables.scss';
-
-.review {
-  display: flex;
-  justify-content: center;
-
-  .signin-button {
-    margin-top: 20px;
-  }
-
-  .review-pane {
-    padding: 10px 40px 20px 40px;
-    width: 100%;
-  }
-
-  h1 {
-    font-size: $size-2;
-    font-weight: $title-weight;
-    color: $title-color;
-  }
-
-  h3.subtitle {
-    color: $primary;
-    font-weight: $weight-normal;
-  }
-
-  .task-analysis-table {
-    margin-bottom: 20px;
-  }
-
-  &__manage {
-    .status-section {
-      .columns {
-        margin-bottom: 0;
-        margin-left: -0.25em;
-
-        .column {
-          padding: 0.25em;
-        }
-      }
-
-      .status-label {
-        font-weight: $weight-light;
-        color: $grey;
-        font-size: $size-6-plus;
-      }
-
-      .status-value {
-        color: $grey;
-        font-size: $size-6-plus;
-      }
-    }
-
-    &__controls.columns {
-      font-size: $size-5;
-      display: flex;
-      justify-content: flex-end;
-      align-items: center;
-      margin: 0;
-
-      .admin__manage__controls--control {
-        padding: 8px 12px;
-
-        a {
-          color: #91F3A2;
-
-          .basic-dropdown-indicator {
-            border-color: #91F3A2;
-            vertical-align: super;
-          }
-
-          &:hover {
-            color: $white;
-
-            .basic-dropdown-indicator {
-              border-color: $white;
-            }
-          }
-        }
-
-        svg {
-          fill: #91F3A2;
-          height: $size-5;
-          width: auto;
-
-          &:hover {
-            fill: $white;
-          }
-        }
-
-        &:first-child {
-          padding-left: 0;
-        }
-      }
-    }
-
-    &__pane-wrapper {
-      display: flex;
-      justify-content: center;
-    }
-
-    &__primary-content {
-      padding: 15px 0 15px 15px;
-      flex: 1 2 600px;
-      min-width: 350px;
-      max-width: 1000px;
-      position: relative;
-
-      &.wide {
-        max-width: inherit;
-      }
-    }
-
-    &__managed-item-list {
-      font-size: $size-5;
-
-      svg.icon {
-        height: 18px;
-        width: auto;
-      }
-
-      .item-status-icon svg.icon {
-        height: 20px;
-        width: auto;
-        vertical-align: middle;
-      }
-
-      .item-entry {
-        &:not(:last-child) {
-          margin-bottom: 0.75rem;
-        }
-
-        .columns {
-          margin: 0;
-
-          .column a {
-            vertical-align: middle;
-          }
-        }
-
-        .column.item-visibility {
-          max-width: 40px;
-
-          svg.icon {
-            height: 24px;
-            width: auto;
-            fill: $green-light;
-            vertical-align: middle;
-
-            &.enabled {
-              fill: $green;
-            }
-          }
-        }
-
-        .column.item-pinned {
-          max-width: 50px;
-
-          svg.icon {
-            height: 18px;
-            width: auto;
-            fill: $grey-lightest;
-            vertical-align: middle;
-            transform: rotate(10deg);
-
-            &.enabled {
-              fill: $blue;
-            }
-          }
-        }
-
-        .column.item-id {
-          max-width: 100px;
-        }
-
-        .column.item-priority {
-          max-width: 100px;
-        }
-
-        .column.controls {
-          color: $grey-light;
-
-          svg {
-            fill: $grey-light;
-
-            &:hover {
-              fill: $grey-darkest;
-            }
-          }
-
-          a, a:link, a:visited, a:active {
-            color: $grey-light;
-          }
-
-          a:hover {
-            color: $grey-darkest;
-          }
-        }
-      }
-    }
-  }
-}
